requirements.txt
flask
groq
python-dotenv
azure-storage-blob
email-validator